Uncomment the comment out code due to Wills input about type conversion

+++ src/main/java/uk/gov/hmcts/appregister/applicationentry/dto/ApplicationWriteDto.java
@@ -9,7 +9,7 @@ public record ApplicationWriteDto(
         Long applicationCodeId,
         IdentityDetailsWriteDto applicant,
         IdentityDetailsWriteDto respondent,
-        Integer numberOfBulkRespondents,
+        Short numberOfBulkRespondents,
         List<String> textFields,
         String caseReference,
         String accountNumber,
